rename module
not only doing uploads anymore
conversion looks like a ok name
